+++ packages/wrapper/test/helper.ts
@@ -3,7 +3,7 @@
  * Licensed under the MIT License. See LICENSE in the package root for license information.
  * ------------------------------------------------------------------------------------------ */
 
-import { UserConfig, EditorAppType, EditorAppExtended, MonacoEditorLanguageClientWrapper, EditorAppClassic } from 'monaco-editor-wrapper';
+import { UserConfig, EditorAppType } from 'monaco-editor-wrapper';
 
 export const createMonacoEditorDiv = () => {
     const div = document.createElement('div');
@@ -34,31 +34,6 @@ export const createEditorAppConfig = (type: EditorAppType) => {
         },
         useDiffEditor: false,
     };
-};
-
-export const updateExtendedAppPrototyp = async () => {
-    EditorAppExtended.prototype.specifyServices = async () => {
-        console.log('Using overriden EditorAppExtended.prototype.specifyServices');
-        return Promise.resolve({});
-    };
-    EditorAppClassic.prototype.specifyServices = async () => {
-        console.log('Using overriden EditorAppClassic.prototype.specifyServices');
-        return Promise.resolve({});
-    };
-};
-
-/**
- * WA: Create an instance for testing that does not specify any additional services.
- *
- * Prevents:
- * Error: Unable to load extension-file://vscode.theme-defaults/themes/light_modern.json:
- * Unable to read file 'extension-file://vscode.theme-defaults/themes/light_modern.json' (TypeError: Failed to fetch)
- */
-export const createAndInitWrapper = async (userConfig: UserConfig) => {
-    updateExtendedAppPrototyp();
-    const wrapper = new MonacoEditorLanguageClientWrapper();
-    await wrapper.init(userConfig);
-    return wrapper;
 };
 
 /**